powerbi_connector/requirements.txt
powerbiclient==3.1.1
msal==1.24.1
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0